Fixed many small bad practices and bugs (FindBugs).

+++ core/src/main/java/nl/inl/blacklab/index/DownloadCache.java
@@ -208,7 +208,7 @@ public class DownloadCache {
     }
 
     public static void setMaxFileSizeMegs(int maxDownloadedFileSizeMegs) {
-        maxDownloadedFileSize = maxDownloadedFileSizeMegs * 1000000;
+        maxDownloadedFileSize = (long)maxDownloadedFileSizeMegs * 1000000L;
         if (maxDownloadFolderSize < maxDownloadedFileSize)
             maxDownloadedFileSize = maxDownloadFolderSize;
     }
